Refactor JSON encoder to separate file
For better separation of logic between the reusable JSON encode/decode
logic and the users-related logic.
